Restructure CLI: cli.commands.utils
