+++ pacco/cli/commands/binary.py
@@ -6,7 +6,7 @@ import shutil
 from pathlib import Path
 from typing import Dict
 
-from pacco.cli.command_abstract import CommandAbstract
+from pacco.cli.commands.utils.command_abstract import CommandAbstract
 from pacco.manager.utils.cache import Cache
