+++ pacco/cli/commands/pacco.py
@@ -1,6 +1,6 @@
-from pacco.cli.command_abstract import CommandAbstract
+from pacco.cli.commands.utils.command_abstract import CommandAbstract
 from pacco.cli.commands import remote, registry, binary
-from pacco.cli.output_stream import OutputStream
+from pacco.cli.commands.utils.output_stream import OutputStream
 from pacco.manager.remote_manager import RemoteManager
