+++ pacco/cli/commands/utils/output_stream.py
